Created simplified loop for parsing logo quantities
Using a powershell class I constructed a template of logo sizes that are commonly used.
In that class initialization I create the array with all the sizes to reference later.
starting on line 147 I use the array to initialize a loop that goes through and gets the logo counts from the json file and fills the array of objects with the counts.
Immediately following I have another loop that takes the same array and writes the logo counts into the console and the print log for the day. These two loops can be merged in the future.

The missing pick and pack forms error catching should still be fixed. it is uncommented now.